refactor: removed paymentService to align with rest of architecture
